Extract CP869 to its own file
